fastapi/GLOBAL/ABSapi-master/back/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-fastapi-jwt-auth==0.6.6
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==3.4.8
dnspython==2.6.1
fastapi==0.115.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
motor==3.6.0
multidict==6.1.0
numpy==2.1.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
yarl==1.15.0
Jinja2==3.1.5
openpyxl==3.1.5
pytest==8.3.5
